Refactor Player Structure, PlayerHealth Done
