+++ Assets/Prefabs/Character.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: bf66523139bb47ea38a5486fc2f6ac05
+guid: 0a26470ea49caf5c59a66aa0e5f74232
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
