+++ Assets/Prefabs/Environment.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: bf66523139bb47ea38a5486fc2f6ac05
+guid: 0bd6ba82858c8d8dbac2ba38cca1a6f1
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
