+++ Assets/Prefabs/Property.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: bf66523139bb47ea38a5486fc2f6ac05
+guid: d1c40fb92d4d2bb98834bf39f19620ca
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
