+++ Assets/Prefabs/Property/HealthManager.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: bf66523139bb47ea38a5486fc2f6ac05
-DefaultImporter:
+guid: bc5ffed1c37267e01a76672eded53669
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
