+++ Assets/Scenes/Scene1.unity
@@ -417,52 +417,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!1 &2002179584
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 2002179586}
-  - component: {fileID: 2002179585}
-  m_Layer: 0
-  m_Name: HealthManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &2002179585
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2002179584}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a619d2680b83c5f20871bb4cb73c13a9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  healthBar: {fileID: 1807976588}
-  healthAmount: 100
---- !u!4 &2002179586
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2002179584}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &883553326031232862
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -622,15 +576,26 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 5956104239326810478, guid: f0c07222a3b254e5b98a6934114ffd86, type: 3}
   m_PrefabInstance: {fileID: 6980765975552827267}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &3227322621682519769 stripped
+--- !u!114 &3227322621682519776 stripped
 MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1837199189108788406, guid: f0c07222a3b254e5b98a6934114ffd86, type: 3}
+  m_CorrespondingSourceObject: {fileID: 659515783135043062, guid: f0c07222a3b254e5b98a6934114ffd86, type: 3}
   m_PrefabInstance: {fileID: 6980765975552827267}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2f2e9d10fef0f6519ba4581b3c55dbac, type: 3}
+  m_Script: {fileID: 11500000, guid: 0843e973d42eeeb958d6c401fc725aab, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &3227322621682519777 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 2975997888866370563, guid: f0c07222a3b254e5b98a6934114ffd86, type: 3}
+  m_PrefabInstance: {fileID: 6980765975552827267}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bc4d5c529fbfd581f895daef4b4f41c1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1001 &4917478968576407985
@@ -656,7 +621,7 @@ PrefabInstance:
     - target: {fileID: 5240095690223936181, guid: c7349d105a6325eceb75f965fe2d78f2, type: 3}
       propertyPath: playerMovementScript
       value: 
-      objectReference: {fileID: 3227322621682519769}
+      objectReference: {fileID: 3227322621682519777}
     - target: {fileID: 5667891049614619252, guid: c7349d105a6325eceb75f965fe2d78f2, type: 3}
       propertyPath: m_LocalPosition.x
       value: 14.5
@@ -759,6 +724,71 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: f0c07222a3b254e5b98a6934114ffd86, type: 3}
+--- !u!1001 &7235075490459300785
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 351773790169934582, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: healthBar
+      value: 
+      objectReference: {fileID: 1807976588}
+    - target: {fileID: 351773790169934582, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: playerHealth
+      value: 
+      objectReference: {fileID: 3227322621682519776}
+    - target: {fileID: 640006400938320947, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_Name
+      value: HealthManager
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6925494860578652395, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: bc5ffed1c37267e01a76672eded53669, type: 3}
 --- !u!1001 &7612379777034588662
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -873,4 +903,4 @@ SceneRoots:
   - {fileID: 1084120587}
   - {fileID: 4917478968576407985}
   - {fileID: 883553326031232862}
-  - {fileID: 2002179586}
+  - {fileID: 7235075490459300785}
